+++ PanicARDemo/PanicARDemo/PARGLViewController.mm
@@ -132,6 +132,9 @@ static float _orientationAngle;
             [self setOrientationAngle:90];
             break;
         case UIInterfaceOrientationPortrait:
+#if defined(__IPHONE_8_0)
+        case UIInterfaceOrientationUnknown:
+#endif
             [self setOrientationAngle:0];
             break;
     }
@@ -225,7 +228,8 @@ static float _orientationAngle;
     }
 
     // Update the projection matrix
-    self.effect.transform.projectionMatrix = tempProjectionMatrix;}
+    self.effect.transform.projectionMatrix = tempProjectionMatrix;
+}
 
 #pragma mark - touches